requirements.txt
streamlit
PyPDF2
python-dotenv
langchain
langchain-groq
